update the test case and simply the main file
